requirements.txt
langchain
langchain-google-genai
google-api-python-client
python-dotenv
requests